remove canton-network SV deployment and split-SV migration IaC
[static]
