--- cluster/pulumi/common-sv/src/sv.ts
@@ -1,832 +1,0 @@
-// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-import * as postgres from '@canton-network/splice-pulumi-common/src/postgres';
-import * as k8s from '@pulumi/kubernetes';
-import * as pulumi from '@pulumi/pulumi';
-import {
-  activeVersion,
-  ansDomainPrefix,
-  appsAffinityAndTolerations,
-  Auth0Client,
-  btoa,
-  ChartValues,
-  CLUSTER_BASENAME,
-  CLUSTER_HOSTNAME,
-  CnInput,
-  config as envConfig,
-  daContactPoint,
-  DecentralizedSynchronizerMigrationConfig,
-  DecentralizedSynchronizerUpgradeConfig,
-  ExactNamespace,
-  envoyClientIpHeaderEnvVar,
-  failOnAppVersionMismatch,
-  fetchAndInstallParticipantBootstrapDump,
-  getAdditionalJvmOptions,
-  getSvAppApiAudience,
-  imagePullSecret,
-  initialPackageConfigJson,
-  initialSynchronizerFeesConfig,
-  InstalledHelmChart,
-  installSpliceHelmChart,
-  installSvAppSecrets,
-  installValidatorOnboardingSecret,
-  isDevNet,
-  networkWideConfig,
-  participantBootstrapDumpSecretName,
-  PersistenceConfig,
-  persistentHeapDumpsPvc,
-  sanitizedForPostgres,
-  spliceInstanceNames,
-  SplicePostgresConfig,
-  svCometBftGovernanceKeyFromSecret,
-  svCometBftGovernanceKeySecret,
-  SvIdKey,
-  svKeyFromSecret,
-  svOnboardingPollingInterval,
-  svValidatorTopupConfig,
-  svUserIds,
-  validatorOnboardingSecretName,
-} from '@canton-network/splice-pulumi-common';
-import {
-  approvedSvIdentities,
-  CantonBftSynchronizerNode,
-  configForSv,
-  coreSvsToDeploy,
-  DecentralizedSynchronizerNode,
-  initialRound,
-  installScanBulkStorage,
-  installSvLoopback,
-  SingleSvConfiguration,
-  StaticSvConfig,
-  SynchronizerNodes,
-  valuesForSvApp,
-  valuesForSvValidatorApp,
-} from '@canton-network/splice-pulumi-common-sv';
-import { SvConfig, svsConfig } from '@canton-network/splice-pulumi-common-sv/src/config';
-import { readBackupConfig } from '@canton-network/splice-pulumi-common-validator/src/backup';
-import { installValidatorApp } from '@canton-network/splice-pulumi-common-validator/src/validator';
-import {
-  mustInstallSplitwell,
-  mustInstallValidator1,
-  splitwellOnboarding,
-  standaloneValidatorOnboarding,
-  validator1Onboarding,
-} from '@canton-network/splice-pulumi-common-validator/src/validators';
-import {
-  delegatelessAutomationExpectedTaskDuration,
-  delegatelessAutomationExpiredRewardCouponBatchSize,
-  delegatelessAutomationExpiredRewardCouponNumBatches,
-} from '@canton-network/splice-pulumi-common/src/automation';
-import {
-  BucketConfig,
-  installBucketSecret,
-} from '@canton-network/splice-pulumi-common/src/buckets';
-import { spliceConfig } from '@canton-network/splice-pulumi-common/src/config/config';
-import { SplitPostgresInstances } from '@canton-network/splice-pulumi-common/src/config/configs';
-import { initialAmuletPrice } from '@canton-network/splice-pulumi-common/src/initialAmuletPrice';
-import { Postgres } from '@canton-network/splice-pulumi-common/src/postgres';
-import { installRateLimits } from '@canton-network/splice-pulumi-common/src/ratelimit/rateLimit';
-import { topologySnapshotConfig } from '@canton-network/splice-pulumi-common/src/topology-snapshot';
-import { Resource } from '@pulumi/pulumi';
-import pick from 'lodash/pick';
-
-import { installInfo } from './info';
-
-// TODO(#6719) once all clusters have been migrated move the whole module to the sv project
-export async function installSvNodeStandalone(
-  xns: ExactNamespace,
-  staticConfig: StaticSvConfig,
-  config: SingleSvConfiguration,
-  auth0Client: Auth0Client,
-  extraDependsOn: CnInput<Resource>[] = [],
-  // TODO(#6719) once all clusters have been migrated remove this
-  migrationArgs?: MigrationArgs
-): Promise<InstalledSv | undefined> {
-  const nodeName = staticConfig.nodeName;
-  const [sv1StaticConfig, ...otherSvsStaticConfigs] = coreSvsToDeploy;
-  const isFoundingSv = nodeName === sv1StaticConfig.nodeName;
-  const disableOnboardingParticipantPromotionDelay = envConfig.envFlag(
-    'DISABLE_ONBOARDING_PARTICIPANT_PROMOTION_DELAY',
-    false
-  );
-  return await installSvNode(
-    xns,
-    {
-      isFirstSv: isFoundingSv,
-      ...pick(staticConfig, [
-        'nodeName',
-        'ingressName',
-        'onboardingName',
-        'cometBft',
-        'validatorWalletUser',
-        'auth0ValidatorAppName',
-        'auth0SvAppName',
-        'sweep',
-      ]),
-      nodeConfigs: {
-        sv1: {
-          ...sv1StaticConfig.cometBft,
-          ...pick(sv1StaticConfig, ['nodeName', 'ingressName']),
-        },
-        peers: otherSvsStaticConfigs
-          .filter(config => config.nodeName !== nodeName)
-          .map(config => ({
-            ...config.cometBft,
-            ...pick(config, ['nodeName', 'ingressName']),
-          })),
-      },
-      onboarding: isFoundingSv
-        ? {
-            type: 'found-dso',
-            sv1SvRewardWeightBps:
-              approvedSvIdentities().find(
-                identity => identity.name == sv1StaticConfig.onboardingName
-              )?.rewardWeightBps ?? 10_000,
-            roundZeroDuration: envConfig.optionalEnv('ROUND_ZERO_DURATION'),
-            initialRound: initialRound?.toString(),
-          }
-        : {
-            type: 'join-with-key',
-            sponsorApiUrl: `http://sv-app.sv-1:5014`,
-            sponsorScanUrl: `http://scan-app.sv-1:5012`,
-            keys: svKeyFromSecret(
-              staticConfig.svIdKeySecretName ?? `${nodeName.replaceAll('-', '')}-id`
-            ),
-          },
-      auth0Client,
-      expectedValidatorOnboardings: isFoundingSv
-        ? [
-            ...(function* () {
-              if (mustInstallSplitwell) {
-                yield splitwellOnboarding;
-              }
-              if (mustInstallValidator1) {
-                yield validator1Onboarding;
-              }
-              if (standaloneValidatorOnboarding !== undefined) {
-                yield standaloneValidatorOnboarding;
-              }
-            })(),
-          ]
-        : [],
-      isDevNet,
-      ...(await readBackupConfig()),
-      topupConfig: svValidatorTopupConfig,
-      splitPostgresInstances: SplitPostgresInstances,
-      disableOnboardingParticipantPromotionDelay,
-      onboardingPollingInterval: svOnboardingPollingInterval,
-      cometBftGovernanceKey:
-        config.participant?.kms !== undefined
-          ? svCometBftGovernanceKeyFromSecret(
-              staticConfig.cometBftGovernanceKeySecretName ??
-                `${nodeName.replaceAll('-', '')}-cometbft-governance-key`
-            )
-          : undefined,
-      initialRound: initialRound?.toString(),
-      version: config.versionOverride ?? activeVersion,
-      ...config,
-    },
-    DecentralizedSynchronizerUpgradeConfig,
-    extraDependsOn,
-    migrationArgs
-  );
-}
-
-export function installSvKeySecret(
-  xns: ExactNamespace,
-  keys: CnInput<SvIdKey>
-): k8s.core.v1.Secret[] {
-  const legacySecretName = 'cn-app-sv-key';
-  const secretName = 'splice-app-sv-key';
-
-  const data = pulumi.output(keys).apply(ks => {
-    return {
-      public: btoa(ks.publicKey),
-      private: btoa(ks.privateKey),
-    };
-  });
-
-  return [
-    new k8s.core.v1.Secret(
-      `cn-app-${xns.logicalName}-key`,
-      {
-        metadata: {
-          name: legacySecretName,
-          namespace: xns.logicalName,
-        },
-        type: 'Opaque',
-        data: data,
-      },
-      {
-        dependsOn: [xns.ns],
-      }
-    ),
-    new k8s.core.v1.Secret(
-      `splice-app-${xns.logicalName}-key`,
-      {
-        metadata: {
-          name: secretName,
-          namespace: xns.logicalName,
-        },
-        type: 'Opaque',
-        data: data,
-      },
-      {
-        dependsOn: [xns.ns],
-      }
-    ),
-  ];
-}
-
-export type InstalledSv = {
-  namespace: ExactNamespace;
-  nodeName: string;
-  validatorApp: Resource;
-  svApp: InstalledHelmChart;
-  scan: InstalledHelmChart;
-  canton: SynchronizerNodes;
-  ingress: Resource;
-  appsPostgres: postgres.Postgres;
-};
-
-export async function installSvNode(
-  xns: ExactNamespace,
-  baseConfig: SvConfig,
-  decentralizedSynchronizerUpgradeConfig: DecentralizedSynchronizerMigrationConfig,
-  extraDependsOn: CnInput<Resource>[] = [],
-  migrationArgs?: MigrationArgs
-): Promise<InstalledSv | undefined> {
-  const periodicBackupConfig: BucketConfig | undefined = baseConfig.periodicBackupConfig
-    ? {
-        ...baseConfig.periodicBackupConfig,
-        location: {
-          ...baseConfig.periodicBackupConfig.location,
-          prefix:
-            baseConfig.periodicBackupConfig.location.prefix ||
-            `${CLUSTER_BASENAME}/${xns.logicalName}`,
-        },
-      }
-    : undefined;
-
-  const svConfig = configForSv(baseConfig.nodeName);
-  const periodicTopologySnapshotConfig = svConfig.periodicSnapshots?.topology
-    ? await topologySnapshotConfig(
-        svConfig.periodicSnapshots?.topology,
-        `${CLUSTER_BASENAME}/${xns.logicalName}`
-      )
-    : undefined;
-
-  const identitiesBackupLocation = {
-    ...baseConfig.identitiesBackupLocation,
-    prefix: baseConfig.identitiesBackupLocation.prefix || `${CLUSTER_BASENAME}/${xns.logicalName}`,
-  };
-
-  const bulkStorageBuckets = svConfig.scanApp?.bulkStorage
-    ? installScanBulkStorage(xns, svConfig.scanApp.bulkStorage)
-    : undefined;
-
-  const config = {
-    ...baseConfig,
-    periodicBackupConfig,
-    identitiesBackupLocation,
-    bulkStorageBuckets,
-  };
-
-  if (migrationArgs?.action === 'import') {
-    const appsPostgres = await installAppsPostgres(xns, config, migrationArgs);
-
-    return undefined;
-  } else {
-    const loopback = installSvLoopback(xns, decentralizedSynchronizerUpgradeConfig.usesCometbft());
-    const imagePullDeps = imagePullSecret(xns);
-
-    const auth0Secrets: CnInput<pulumi.Resource>[] = await installSvAppSecrets(
-      xns,
-      baseConfig.auth0Client
-    );
-
-    const identitiesBackupConfigSecret = installBucketSecret(
-      xns,
-      config.identitiesBackupLocation.bucket
-    );
-
-    const topologySnapshotConfigSecret = periodicTopologySnapshotConfig
-      ? installBucketSecret(xns, periodicTopologySnapshotConfig.location.bucket)
-      : undefined;
-    const backupConfigSecret: pulumi.Resource | undefined = config.periodicBackupConfig
-      ? config.periodicBackupConfig.location.bucket != config.identitiesBackupLocation.bucket
-        ? installBucketSecret(xns, config.periodicBackupConfig.location.bucket)
-        : identitiesBackupConfigSecret
-      : undefined;
-
-    const participantBootstrapDumpSecret: pulumi.Resource | undefined =
-      config.bootstrappingDumpConfig
-        ? await fetchAndInstallParticipantBootstrapDump(xns, config.bootstrappingDumpConfig)
-        : undefined;
-
-    const dependsOn: CnInput<pulumi.Resource>[] = auth0Secrets
-      .concat(
-        config.onboarding.type == 'join-with-key'
-          ? installSvKeySecret(xns, config.onboarding.keys)
-          : []
-      )
-      .concat(
-        config.onboarding.type == 'join-with-key' &&
-          config.onboarding.sponsorRelease !== undefined &&
-          spliceConfig.pulumiProjectConfig.interAppsDependencies
-          ? [config.onboarding.sponsorRelease]
-          : []
-      )
-      .concat(
-        config.expectedValidatorOnboardings.map(onboarding =>
-          installValidatorOnboardingSecret(xns, onboarding.name, onboarding.secret)
-        )
-      )
-      .concat([identitiesBackupConfigSecret])
-      .concat(backupConfigSecret ? [backupConfigSecret] : [])
-      .concat(topologySnapshotConfigSecret ? [topologySnapshotConfigSecret] : [])
-      .concat(participantBootstrapDumpSecret ? [participantBootstrapDumpSecret] : [])
-      .concat(loopback)
-      .concat(imagePullDeps)
-      .concat(
-        config.cometBftGovernanceKey
-          ? svCometBftGovernanceKeySecret(xns, config.cometBftGovernanceKey)
-          : []
-      )
-      .concat(
-        bulkStorageBuckets
-          ? [
-              bulkStorageBuckets.staging.secret,
-              bulkStorageBuckets.staging.bucket,
-              bulkStorageBuckets.committed.secret,
-              bulkStorageBuckets.committed.bucket,
-            ]
-          : []
-      )
-      .concat(extraDependsOn);
-
-    const appsPostgres = await installAppsPostgres(xns, config, migrationArgs);
-
-    const canton = new SynchronizerNodes(
-      decentralizedSynchronizerUpgradeConfig,
-      {
-        ...config.nodeConfigs,
-        self: { ...config.cometBft, nodeName: config.nodeName },
-      },
-      config.ingressName
-    );
-
-    const svApp = installSvApp(
-      decentralizedSynchronizerUpgradeConfig,
-      { ...config, periodicTopologySnapshotConfig },
-      xns,
-      dependsOn,
-      appsPostgres,
-      canton
-    );
-
-    const scan = installScan(
-      xns,
-      config,
-      decentralizedSynchronizerUpgradeConfig,
-      dependsOn,
-      canton,
-      svApp,
-      appsPostgres
-    );
-
-    installInfo(
-      xns,
-      `info.${config.ingressName}.${CLUSTER_HOSTNAME}`,
-      'cluster-ingress/cn-http-gateway',
-      decentralizedSynchronizerUpgradeConfig,
-      `http://scan-app.${config.nodeName}:5012`,
-      scan,
-      config.version
-    );
-
-    const validatorApp = await installValidator(
-      appsPostgres,
-      xns,
-      decentralizedSynchronizerUpgradeConfig,
-      baseConfig,
-      backupConfigSecret,
-      canton,
-      svApp,
-      scan
-    );
-
-    const ingress = installSpliceHelmChart(
-      xns,
-      'ingress-sv',
-      'splice-cluster-ingress-runbook',
-      {
-        withSvIngress: true,
-        ingress: {
-          decentralizedSynchronizer: {
-            migrationIds: decentralizedSynchronizerUpgradeConfig
-              .runningMigrations()
-              .map(x => x.id.toString()),
-          },
-        },
-        spliceDomainNames: {
-          nameServiceDomain: ansDomainPrefix,
-        },
-        cluster: {
-          hostname: CLUSTER_HOSTNAME,
-          svNamespace: xns.logicalName,
-          svIngressName: config.ingressName,
-        },
-        rateLimit: {
-          scan: {
-            enable: false,
-          },
-        },
-      },
-      config.version,
-      { dependsOn: [xns.ns] }
-    );
-
-    return {
-      namespace: xns,
-      nodeName: config.nodeName,
-      canton,
-      validatorApp,
-      svApp,
-      scan,
-      ingress,
-      appsPostgres,
-    };
-  }
-}
-
-async function installAppsPostgres(
-  xns: ExactNamespace,
-  config: SvConfig,
-  migrationArgs?: MigrationArgs
-): Promise<postgres.Postgres> {
-  const defaultPostgres = config.splitPostgresInstances
-    ? undefined
-    : await postgres.installPostgres(
-        xns,
-        'postgres',
-        'postgres',
-        config.version,
-        spliceConfig.pulumiProjectConfig.cloudSql,
-        spliceConfig.pulumiProjectConfig.defaultSplicePostgresConfig,
-        false,
-        {
-          logicalDecoding: !!config.scanApp?.bigQuery,
-        }
-      );
-
-  const appsPostgres =
-    defaultPostgres ||
-    (await postgres.installPostgres(
-      xns,
-      `cn-apps-pg`,
-      `cn-apps-pg`,
-      config.version,
-      config.appsPg?.cloudSql ?? spliceConfig.pulumiProjectConfig.cloudSql,
-      spliceConfig.pulumiProjectConfig.defaultSplicePostgresConfig,
-      true,
-      {
-        logicalDecoding: !!config.scanApp?.bigQuery,
-        ...(() => {
-          switch (migrationArgs?.action) {
-            case 'import':
-              return {
-                existingInstanceName: migrationArgs.databaseInstanceName,
-                existingSecretName: migrationArgs.databaseSecretName,
-              };
-            case 'export':
-              return {
-                retainDbResourcesOnDelete: true,
-              };
-            case undefined:
-              return {};
-          }
-        })(),
-      }
-    ));
-
-  return appsPostgres;
-}
-
-function persistenceConfig(postgresDb: postgres.Postgres, dbName: string): PersistenceConfig {
-  const dbNameO = pulumi.Output.create(dbName);
-  return {
-    host: postgresDb.address,
-    databaseName: dbNameO,
-    secretName: postgresDb.secretName,
-    schema: dbNameO,
-    user: pulumi.Output.create('cnadmin'),
-    port: pulumi.Output.create(5432),
-    postgresName: postgresDb.instanceName,
-  };
-}
-
-async function installValidator(
-  postgres: Postgres,
-  xns: ExactNamespace,
-  decentralizedSynchronizerMigrationConfig: DecentralizedSynchronizerMigrationConfig,
-  svConfig: SvConfig,
-  backupConfigSecret: Resource | undefined,
-  sv: SynchronizerNodes,
-  svApp: Resource,
-  scan: Resource
-) {
-  const validatorDbName = `validator_${sanitizedForPostgres(svConfig.nodeName)}`;
-  const commonValidatorAppValues = valuesForSvValidatorApp(
-    decentralizedSynchronizerMigrationConfig,
-    svConfig
-  );
-
-  return await installValidatorApp({
-    xns,
-    ...commonValidatorAppValues,
-    validatorWalletUsers: svUserIds(svConfig.auth0Client.getCfg()).apply(ids =>
-      ids.concat(svConfig.validatorWalletUser ? [svConfig.validatorWalletUser] : [])
-    ),
-    dependencies: [],
-    disableAllocateLedgerApiUserParty: true,
-    topupConfig: svConfig.topupConfig,
-    backupConfig:
-      svConfig.periodicBackupConfig && backupConfigSecret
-        ? {
-            config: svConfig.periodicBackupConfig,
-            secret: backupConfigSecret,
-          }
-        : undefined,
-    persistenceConfig: persistenceConfig(postgres, validatorDbName),
-    extraDependsOn: spliceConfig.pulumiProjectConfig.interAppsDependencies
-      ? [svApp, postgres, scan]
-      : [postgres],
-    svValidator: true,
-    participantAddress: sv.participant.internalClusterAddress,
-    scanAddress: internalScanUrl(svConfig),
-    auth0Client: svConfig.auth0Client,
-    auth0ValidatorAppName: svConfig.auth0ValidatorAppName,
-    sweep: svConfig.sweep,
-    nodeIdentifier: svConfig.onboardingName,
-    logLevel: svConfig.logging?.appsLogLevel,
-    logAsync: svConfig.logging?.appsAsync,
-    apiRequestLogLevel: svConfig.logging?.apiRequestLogLevel,
-    additionalJvmOptions: svConfig.validatorApp?.additionalJvmOptions || '',
-    resources: svConfig.validatorApp?.resources,
-    version: svConfig.version,
-  });
-}
-
-function publicScanUrl(config: SvConfig) {
-  return `https://scan.${config.ingressName}.${CLUSTER_HOSTNAME}`;
-}
-
-function internalScanUrl(config: SvConfig): string {
-  return `http://scan-app.${config.nodeName}:5012`;
-}
-
-function installSvApp(
-  decentralizedSynchronizerMigrationConfig: DecentralizedSynchronizerMigrationConfig,
-  config: SvConfig,
-  xns: ExactNamespace,
-  dependsOn: CnInput<Resource>[],
-  postgres: Postgres,
-  synchronizerNodes: SynchronizerNodes
-) {
-  const { participant } = synchronizerNodes;
-  const decentralizedSynchronizer = synchronizerNodes.active;
-  const allSynchronizerDependencies = [
-    synchronizerNodes.active,
-    synchronizerNodes.legacy,
-    synchronizerNodes.upgrade,
-    ...synchronizerNodes.additionalLegacy,
-  ]
-    .filter((n): n is NonNullable<typeof n> => n !== undefined)
-    .flatMap(n => n.dependencies);
-  const svDbName = `sv_${sanitizedForPostgres(config.nodeName)}`;
-  const commonSvAppValues = valuesForSvApp(
-    decentralizedSynchronizerMigrationConfig,
-    { ...config, skipInitialization: svsConfig?.synchronizer?.skipInitialization },
-    synchronizerNodes,
-    config.ingressName
-  );
-
-  const svValues = {
-    ...commonSvAppValues,
-    ...spliceInstanceNames,
-    onboardingType: config.onboarding.type,
-    onboardingName: config.onboardingName,
-    onboardingFoundingSvRewardWeightBps:
-      config.onboarding.type == 'found-dso' ? config.onboarding.sv1SvRewardWeightBps : undefined,
-    onboardingRoundZeroDuration:
-      config.onboarding.type == 'found-dso' ? config.onboarding.roundZeroDuration : undefined,
-    initialSynchronizerFeesConfig:
-      config.onboarding.type == 'found-dso' ? initialSynchronizerFeesConfig : undefined,
-    initialPackageConfigJson:
-      config.onboarding.type == 'found-dso' ? initialPackageConfigJson : undefined,
-    initialRound:
-      config.onboarding.type == 'found-dso' ? config.onboarding.initialRound : undefined,
-    initialAmuletPrice: initialAmuletPrice,
-    disableOnboardingParticipantPromotionDelay: config.disableOnboardingParticipantPromotionDelay,
-    decentralizedSynchronizerUrl:
-      config.onboarding.type == 'found-dso'
-        ? undefined
-        : decentralizedSynchronizer.sv1InternalSequencerAddress,
-    scan: {
-      publicUrl: publicScanUrl(config),
-      internalUrl: internalScanUrl(config),
-    },
-    expectedValidatorOnboardings: config.expectedValidatorOnboardings.map(onboarding => ({
-      expiresIn: onboarding.expiresIn,
-      secretFrom: {
-        secretKeyRef: {
-          name: validatorOnboardingSecretName(onboarding.name),
-          key: 'secret',
-          optional: false,
-        },
-      },
-    })),
-    isDevNet: config.isDevNet,
-    approvedSvIdentities: approvedSvIdentities(),
-    persistence: persistenceConfig(postgres, svDbName),
-    identitiesExport: config.identitiesBackupLocation,
-    participantIdentitiesDumpImport: config.bootstrappingDumpConfig
-      ? { secretName: participantBootstrapDumpSecretName }
-      : undefined,
-    metrics: {
-      enable: true,
-    },
-    additionalJvmOptions: getAdditionalJvmOptions(config.svApp?.additionalJvmOptions),
-    failOnAppVersionMismatch: failOnAppVersionMismatch,
-    participantAddress: participant.internalClusterAddress,
-    onboardingPollingInterval: config.onboardingPollingInterval,
-    enablePostgresMetrics: true,
-    auth: {
-      audience: getSvAppApiAudience(config.auth0Client.getCfg(), xns.logicalName),
-      jwksUrl: `https://${config.auth0Client.getCfg().auth0Domain}/.well-known/jwks.json`,
-    },
-    contactPoint: daContactPoint,
-    nodeIdentifier: config.onboardingName,
-    delegatelessAutomationExpectedTaskDuration: delegatelessAutomationExpectedTaskDuration,
-    delegatelessAutomationExpiredRewardCouponBatchSize:
-      delegatelessAutomationExpiredRewardCouponBatchSize,
-    delegatelessAutomationExpiredRewardCouponNumBatches:
-      delegatelessAutomationExpiredRewardCouponNumBatches,
-    maxVettingDelay: networkWideConfig?.maxVettingDelay,
-    logLevel: config.logging?.appsLogLevel,
-    apiRequestLogLevel: config.logging?.apiRequestLogLevel,
-    logAsyncFlush: config.logging?.appsAsync,
-    resources: config.svApp?.resources,
-    periodicTopologySnapshotConfig: config.periodicTopologySnapshotConfig,
-    persistentDataPvc: persistentHeapDumpsPvc(),
-  } as ChartValues;
-
-  if (config.onboarding.type == 'join-with-key') {
-    svValues.joinWithKeyOnboarding = {
-      sponsorApiUrl: config.onboarding.sponsorApiUrl,
-    };
-  }
-
-  return installSpliceHelmChart(
-    xns,
-    'sv-app',
-    'splice-sv-node',
-    svValues,
-    config.version,
-    {
-      dependsOn: dependsOn.concat([postgres]).concat(allSynchronizerDependencies),
-    },
-    undefined,
-    appsAffinityAndTolerations
-  );
-}
-
-function installScan(
-  xns: ExactNamespace,
-  config: SvConfig,
-  decentralizedSynchronizerMigrationConfig: DecentralizedSynchronizerMigrationConfig,
-  dependsOn: CnInput<Resource>[],
-  synchronizerNodes: SynchronizerNodes,
-  svApp: pulumi.Resource,
-  postgres: Postgres
-) {
-  const useCantonBft = decentralizedSynchronizerMigrationConfig.active.sequencer.enableBftSequencer;
-  const { active, participant } = synchronizerNodes;
-  const scanDbName = `scan_${sanitizedForPostgres(config.nodeName)}`;
-
-  const cantonBftConfigFor = (node: DecentralizedSynchronizerNode) => {
-    return {
-      cantonBft: {
-        p2pUrl: (node as unknown as CantonBftSynchronizerNode).externalSequencerP2pAddress,
-      },
-    };
-  };
-
-  const synchronizerValues = {
-    synchronizers: {
-      current: {
-        sequencer: active.namespaceInternalSequencerAddress,
-        mediator: active.namespaceInternalMediatorAddress,
-        ...(useCantonBft ? cantonBftConfigFor(active) : {}),
-      },
-      ...(synchronizerNodes.upgrade
-        ? {
-            successor: {
-              sequencer: synchronizerNodes.upgrade.namespaceInternalSequencerAddress,
-              mediator: synchronizerNodes.upgrade.namespaceInternalMediatorAddress,
-              ...(decentralizedSynchronizerMigrationConfig.upgrade?.sequencer.enableBftSequencer
-                ? cantonBftConfigFor(synchronizerNodes.upgrade)
-                : {}),
-            },
-          }
-        : {}),
-      ...(synchronizerNodes.legacy
-        ? {
-            legacy: {
-              sequencer: synchronizerNodes.legacy.namespaceInternalSequencerAddress,
-              mediator: synchronizerNodes.legacy.namespaceInternalMediatorAddress,
-              ...(decentralizedSynchronizerMigrationConfig.legacy?.sequencer.enableBftSequencer
-                ? cantonBftConfigFor(synchronizerNodes.legacy)
-                : {}),
-            },
-          }
-        : {}),
-    },
-  };
-
-  const scanValues = {
-    ...spliceInstanceNames,
-    metrics: {
-      enable: true,
-    },
-    isFirstSv: config.isFirstSv,
-    persistence: persistenceConfig(postgres, scanDbName),
-    additionalJvmOptions: getAdditionalJvmOptions(config.scanApp?.additionalJvmOptions),
-    failOnAppVersionMismatch: failOnAppVersionMismatch,
-    participantAddress: participant.internalClusterAddress,
-    ...(config.onboarding.type == 'join-with-key'
-      ? { sponsorScanUrl: config.onboarding.sponsorScanUrl }
-      : {}),
-    ...synchronizerValues,
-    enablePostgresMetrics: true,
-    logLevel: config.logging?.appsLogLevel,
-    apiRequestLogLevel: config.logging?.apiRequestLogLevel,
-    logAsyncFlush: config.logging?.appsAsync,
-    additionalEnvVars: (config.scanApp?.additionalEnvVars || []).concat([
-      envoyClientIpHeaderEnvVar('canton.scan-apps.scan-app'),
-    ]),
-    resources: config.scanApp?.resources,
-    ...(config.bulkStorageBuckets
-      ? {
-          bulkStorage: {
-            staging: {
-              region: config.bulkStorageBuckets.staging.region,
-              bucketName: config.bulkStorageBuckets.staging.bucket.name,
-              endpoint: 'https://storage.googleapis.com', // gcs endpoint for s3
-              secretName: config.bulkStorageBuckets.staging.secret.metadata.name,
-            },
-            committed: {
-              region: config.bulkStorageBuckets.committed.region,
-              bucketName: config.bulkStorageBuckets.committed.bucket.name,
-              endpoint: 'https://storage.googleapis.com', // gcs endpoint for s3
-              secretName: config.bulkStorageBuckets.committed.secret.metadata.name,
-            },
-          },
-        }
-      : {}),
-    publicUrl: publicScanUrl(config),
-    pvc: persistentHeapDumpsPvc(),
-  };
-
-  if (svsConfig?.scan?.externalRateLimits) {
-    installRateLimits(xns.logicalName, 'scan-app', 5012, svsConfig.scan.externalRateLimits);
-  }
-
-  return installSpliceHelmChart(xns, 'scan', 'splice-scan', scanValues, config.version, {
-    // TODO(#893) if possible, don't require parallel start of sv app and scan when using CantonBft
-    dependsOn: dependsOn
-      .concat(active.dependencies)
-      .concat(
-        spliceConfig.pulumiProjectConfig.interAppsDependencies && !useCantonBft
-          ? [svApp]
-          : [postgres]
-      ),
-  });
-}
-
-export type MigrationArgs =
-  | {
-      action: 'import';
-      databaseInstanceName: string;
-      databaseSecretName: string;
-    }
-  | {
-      action: 'export';
-    };
-
-export type SvsMigrationOutput = Array<{
-  nodeName: string;
-  databaseInstanceName: string;
-  databaseSecretName: string;
-}>;
