+++ cluster/pulumi/sv/src/installNode.ts
@@ -7,7 +7,6 @@ import {
   exactNamespace,
   imagePullSecretWithNonDefaultServiceAccount,
   installLedgerApiUserSecret,
-  spliceConfig,
 } from '@canton-network/splice-pulumi-common';
 import {
   configForSv,
@@ -15,38 +14,21 @@ import {
   svConfigs,
   svRunbookConfig,
 } from '@canton-network/splice-pulumi-common-sv';
-import {
-  installSvNodeStandalone,
-  MigrationArgs,
-  SvsMigrationOutput,
-} from '@canton-network/splice-pulumi-common-sv/src/sv';
-import { StackReferences } from '@canton-network/splice-pulumi-common/src/stackReferences';
 
 import { installParticipant } from './participant';
+import { installSvNode } from './sv';
 
 export async function installNode(sv: string, auth0Client: Auth0Client): Promise<void> {
-  // TODO(#6719) once all clusters have been migrated hardcode splitSvDeploymentEnabled to true
-  const splitSvDeploymentEnabled =
-    spliceConfig.configuration.synchronizerMigration.splitSvDeploymentEnabled;
   const staticConfig = findStaticConfigOrFail(sv);
   const config = configForSv(staticConfig.nodeName);
-  const xns = exactNamespace(staticConfig.nodeName, true, !splitSvDeploymentEnabled);
+  const xns = exactNamespace(staticConfig.nodeName, true, false);
   const serviceAccountName = 'sv';
   const imagePullDeps = imagePullSecretWithNonDefaultServiceAccount(xns, serviceAccountName);
   const auth0Config = auth0Client.getCfg();
   const ledgerApiUserSecret = installLedgerApiUserSecret(auth0Client, xns, 'sv', 'sv');
   const ledgerApiUserSecretSource = auth0UserNameEnvVarSource('sv', true);
-  // TODO(#6719) once all clusters have been migrated remove this
-  const migrateToSplitSvDeployment =
-    spliceConfig.configuration.synchronizerMigration.migrateToSplitSvDeployment;
-  if (
-    (splitSvDeploymentEnabled || migrateToSplitSvDeployment) &&
-    staticConfig.nodeName !== svRunbookConfig.nodeName
-  ) {
-    const migrationArgs = migrateToSplitSvDeployment
-      ? await getMigrationArgsForSv(staticConfig.nodeName)
-      : undefined;
-    await installSvNodeStandalone(xns, staticConfig, config, auth0Client, [], migrationArgs);
+  if (staticConfig.nodeName !== svRunbookConfig.nodeName) {
+    await installSvNode(xns, staticConfig, config, auth0Client, []);
   }
   await installParticipant(
     {
@@ -72,19 +54,4 @@ function findStaticConfigOrFail(sv: string): StaticSvConfig {
   } else {
     return svConfig;
   }
-}
-
-// TODO(#6719) once all clusters have been migrated remove this
-async function getMigrationArgsForSv(nodeName: string): Promise<MigrationArgs> {
-  const svs = (await StackReferences.cantonNetwork.requireOutputValue('svs')) as SvsMigrationOutput;
-  const sv =
-    svs.find(sv => sv.nodeName === nodeName) ??
-    (() => {
-      throw new Error(`No migration output found for SV: ${nodeName}`);
-    })();
-  return {
-    action: 'import',
-    databaseInstanceName: sv.databaseInstanceName,
-    databaseSecretName: sv.databaseSecretName,
-  };
 }
